+++ src/csc229/dealership/Ford_F150.java
@@ -1,21 +1,6 @@
 package csc229.dealership;
 
-public class Ford_F150 {
-
-    private String modelTrim;
-    private int year;
-    private double price;
-    private String tireBrand;
-    private String transmissionType;
-    private int topSpeed;
-    private String entertainmentSystemManufacturer;
-    private String seatStyle;
-    private int modelYear;
-    private double oilCapacity;
-    private double gasTankSize;
-    private double milesPerGallon;
-    private double truckBedLength;
-    private double towingCapacity;
+public class Ford_F150 extends Trucks {
 
     // Constructor
     public Ford_F150(String modelTrim, int year, double price, String tireBrand, String transmissionType,
@@ -38,116 +23,6 @@ public class Ford_F150 {
         this.towingCapacity = towingCapacity;
     }
 
-    // Getters and setters
-    public String getModelTrim() {
-        return modelTrim;
-    }
-
-    public void setModelTrim(String modelTrim) {
-        this.modelTrim = modelTrim;
-    }
-
-    public int getPurchaseYear() {
-        return year;
-    }
-
-    public void setPurchaseYear(int year) {
-        this.year = year;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
-    public String getTireBrand() {
-        return tireBrand;
-    }
-
-    public void setTireBrand(String tireBrand) {
-        this.tireBrand = tireBrand;
-    }
-
-    public String getTransmissionType() {
-        return transmissionType;
-    }
-
-    public void setTransmissionType(String transmissionType) {
-        this.transmissionType = transmissionType;
-    }
-
-    public int getTopSpeed() {
-        return topSpeed;
-    }
-
-    public void setTopSpeed(int topSpeed) {
-        this.topSpeed = topSpeed;
-    }
-
-    public String getEntertainmentSystemManufacturer() {
-        return entertainmentSystemManufacturer;
-    }
-
-    public void setEntertainmentSystemManufacturer(String entertainmentSystemManufacturer) {
-        this.entertainmentSystemManufacturer = entertainmentSystemManufacturer;
-    }
-
-    public String getSeatStyle() {
-        return seatStyle;
-    }
-
-    public void setSeatStyle(String seatStyle) {
-        this.seatStyle = seatStyle;
-    }
-
-    public int getModelYear() {
-        return modelYear;
-    }
-
-    public void setModelYear(int modelYear) {
-        this.modelYear = modelYear;
-    }
-
-    public double getOilCapacity() {
-        return oilCapacity;
-    }
-
-    public void setOilCapacity(double oilCapacity) {
-        this.oilCapacity = oilCapacity;
-    }
-
-    public double getGasTankSize() {
-        return gasTankSize;
-    }
-
-    public void setGasTankSize(double gasTankSize) {
-        this.gasTankSize = gasTankSize;
-    }
-
-    public double getMilesPerGallon() {
-        return milesPerGallon;
-    }
-
-    public void setMilesPerGallon(double milesPerGallon) {
-        this.milesPerGallon = milesPerGallon;
-    }
-    
-    
-    public double getTruckBedLength() {
-        return this.truckBedLength;
-    }
-    public double getTowingCapacity() {
-        return this.towingCapacity;
-    }
-    public void setTruckBedLength(double truckBedLength) {
-        this.truckBedLength = truckBedLength;
-    }
-    public void setTowingCapacity(int towingCapacity) {
-        this.towingCapacity = towingCapacity;
-    }
     
  // toString method
     @Override
