+++ src/csc229/dealership/HarleyDavidson.java
@@ -1,19 +1,6 @@
 package csc229.dealership;
 
-public class HarleyDavidson {
-
-    private String modelTrim;
-    private int year;
-    private double price;
-    private String tireBrand;
-    private String transmissionType;
-    private int topSpeed;
-    private int modelYear;
-    private double oilCapacity;
-    private double gasTankSize;
-    private double milesPerGallon;
-    private String sidecarStyle;
-
+public class HarleyDavidson extends Motorcycle {
     // Constructor
     public HarleyDavidson(String modelTrim, int year, double price, String tireBrand, String transmissionType,
                  int topSpeed, int modelYear, double oilCapacity, double gasTankSize, double milesPerGallon,
@@ -30,95 +17,7 @@ public class HarleyDavidson {
         this.milesPerGallon = milesPerGallon;
         this.sidecarStyle = sidecarStyle;
     }
-
-    // Getters and setters
-    public String getModelTrim() {
-        return modelTrim;
-    }
-
-    public void setModelTrim(String modelTrim) {
-        this.modelTrim = modelTrim;
-    }
-
-    public int getPurchaseYear() {
-        return year;
-    }
-
-    public void setPurchaseYear(int year) {
-        this.year = year;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
-    public String getTireBrand() {
-        return tireBrand;
-    }
-
-    public void setTireBrand(String tireBrand) {
-        this.tireBrand = tireBrand;
-    }
-
-    public String getTransmissionType() {
-        return transmissionType;
-    }
-
-    public void setTransmissionType(String transmissionType) {
-        this.transmissionType = transmissionType;
-    }
-
-    public int getTopSpeed() {
-        return topSpeed;
-    }
-
-    public void setTopSpeed(int topSpeed) {
-        this.topSpeed = topSpeed;
-    }
-
-    public int getModelYear() {
-        return modelYear;
-    }
-
-    public void setModelYear(int modelYear) {
-        this.modelYear = modelYear;
-    }
-
-    public double getOilCapacity() {
-        return oilCapacity;
-    }
-
-    public void setOilCapacity(double oilCapacity) {
-        this.oilCapacity = oilCapacity;
-    }
-
-    public double getGasTankSize() {
-        return gasTankSize;
-    }
-
-    public void setGasTankSize(double gasTankSize) {
-        this.gasTankSize = gasTankSize;
-    }
-
-    public double getMilesPerGallon() {
-        return milesPerGallon;
-    }
-
-    public void setMilesPerGallon(double milesPerGallon) {
-        this.milesPerGallon = milesPerGallon;
-    }
-
-    public String getSidecarStyle() {
-        return sidecarStyle;
-    }
-
-    public void setSidecarStyle(String sidecarStyle) {
-        this.sidecarStyle = sidecarStyle;
-    }
+  
 
  // toString method
     @Override
