+++ src/csc229/dealership/LateModelToyota.java
@@ -1,18 +1,6 @@
 package csc229.dealership;
 
-public class LateModelToyota {
-
-    private String modelTrim;
-    private int year;
-    private double price;
-    private String tireBrand;
-    private String transmissionType;
-    private int topSpeed;
-    private int modelYear;
-    private double oilCapacity;
-    private double gasTankSize;
-    private double milesPerGallon;
-
+public class LateModelToyota extends Vehicles {
     // Constructor
     public LateModelToyota() {
     	
@@ -31,86 +19,6 @@ public class LateModelToyota {
         this.milesPerGallon = milesPerGallon;
     }
 
-    // Getters and setters
-    public String getModelTrim() {
-        return modelTrim;
-    }
-
-    public void setModelTrim(String modelTrim) {
-        this.modelTrim = modelTrim;
-    }
-
-    public int getPurchaseYear() {
-        return year;
-    }
-
-    public void setPurchaseYear(int year) {
-        this.year = year;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
-    public String getTireBrand() {
-        return tireBrand;
-    }
-
-    public void setTireBrand(String tireBrand) {
-        this.tireBrand = tireBrand;
-    }
-
-    public String getTransmissionType() {
-        return transmissionType;
-    }
-
-    public void setTransmissionType(String transmissionType) {
-        this.transmissionType = transmissionType;
-    }
-
-    public int getTopSpeed() {
-        return topSpeed;
-    }
-
-    public void setTopSpeed(int topSpeed) {
-        this.topSpeed = topSpeed;
-    }
-
-    public int getModelYear() {
-        return modelYear;
-    }
-
-    public void setModelYear(int modelYear) {
-        this.modelYear = modelYear;
-    }
-
-    public double getOilCapacity() {
-        return oilCapacity;
-    }
-
-    public void setOilCapacity(double oilCapacity) {
-        this.oilCapacity = oilCapacity;
-    }
-
-    public double getGasTankSize() {
-        return gasTankSize;
-    }
-
-    public void setGasTankSize(double gasTankSize) {
-        this.gasTankSize = gasTankSize;
-    }
-
-    public double getMilesPerGallon() {
-        return milesPerGallon;
-    }
-
-    public void setMilesPerGallon(double milesPerGallon) {
-        this.milesPerGallon = milesPerGallon;
-    }
     @Override
     public String toString() {
     	return "This is a Late Model Toyota -> " +
