+++ src/csc229/dealership/Motorcycle.java
@@ -1,0 +1,15 @@
+package csc229.dealership;
+
+public class Motorcycle  extends Vehicles{
+	protected String sidecarStyle;
+	public String getSidecarStyle() {
+        return sidecarStyle;
+    }
+
+    public void setSidecarStyle(String sidecarStyle) {
+        this.sidecarStyle = sidecarStyle;
+    }
+	
+	
+
+}
