+++ src/csc229/dealership/RoofVehicles.java
@@ -1,0 +1,14 @@
+package csc229.dealership;
+
+public abstract class RoofVehicles extends EnterSystemManu {
+	protected double sunroofSize;
+	
+	   public double getSunroofSize() {
+	        return sunroofSize;
+	    }
+
+	    public void setSunroofSize(double sunroofSize) {
+	        this.sunroofSize = sunroofSize;
+	    }
+
+}
