+++ src/csc229/dealership/Tesla.java
@@ -1,19 +1,6 @@
 package csc229.dealership;
 
-public class Tesla {
-
-    private String modelTrim;
-    private int year;
-    private double price;
-    private String tireBrand;
-    private int topSpeed;
-    private String entertainmentSystemManufacturer;
-    private String seatStyle;
-    private int modelYear;
-    private double sunroofSize;
-    private boolean selfDriving;
-    private double milesPerKilowattHour;
-
+public class Tesla extends EV{
     // Constructor
     public Tesla() {
     	
@@ -33,96 +20,6 @@ public class Tesla {
         this.selfDriving = selfDriving;
         this.milesPerKilowattHour = milesPerKilowattHour;
     }
-
-    // Getters and setters
-    public String getModelTrim() {
-        return modelTrim;
-    }
-
-    public void setModelTrim(String modelTrim) {
-        this.modelTrim = modelTrim;
-    }
-
-    public int getPurchaseYear() {
-        return year;
-    }
-
-    public void setPurchaseYear(int year) {
-        this.year = year;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
-    public String getTireBrand() {
-        return tireBrand;
-    }
-
-    public void setTireBrand(String tireBrand) {
-        this.tireBrand = tireBrand;
-    }
-
-    public int getTopSpeed() {
-        return topSpeed;
-    }
-
-    public void setTopSpeed(int topSpeed) {
-        this.topSpeed = topSpeed;
-    }
-
-    public String getEntertainmentSystemManufacturer() {
-        return entertainmentSystemManufacturer;
-    }
-
-    public void setEntertainmentSystemManufacturer(String entertainmentSystemManufacturer) {
-        this.entertainmentSystemManufacturer = entertainmentSystemManufacturer;
-    }
-
-    public String getSeatStyle() {
-        return seatStyle;
-    }
-
-    public void setSeatStyle(String seatStyle) {
-        this.seatStyle = seatStyle;
-    }
-
-    public int getModelYear() {
-        return modelYear;
-    }
-
-    public void setModelYear(int modelYear) {
-        this.modelYear = modelYear;
-    }
-
-    public double getSunroofSize() {
-        return sunroofSize;
-    }
-
-    public void setSunroofSize(double sunroofSize) {
-        this.sunroofSize = sunroofSize;
-    }
-
-    public boolean isSelfDriving() {
-        return selfDriving;
-    }
-
-    public void setSelfDriving(boolean selfDriving) {
-        this.selfDriving = selfDriving;
-    }
-
-    public double getMilesPerKilowattHour() {
-        return milesPerKilowattHour;
-    }
-
-    public void setMilesPerKilowattHour(double milesPerKilowattHour) {
-        this.milesPerKilowattHour = milesPerKilowattHour;
-    }
-
  // toString method
     @Override
     public String toString() {
